Add suppressing newline character flag
The newline character (0x0A) is appended to the output when running
interactively by reading in STDIN. This pretties up the output for the
interactive session, but it can cause problems when piping the output to
other applications that read in binary data or do not use the newline
character for message/line termination. Since there is no way to know if
the input is coming from a redirections, pipe, or interactively, this
flag was added to suppress the newline character if specified.