Change input reading to bytes instead of strings
The reading of input, either from STDIN or a file, is the act of reading
bytes, not necessarily strings. The input could be a binary format, such
as MessagePack (not JSON). The safer and more generic implementation is
to treat reading input as reading bytes, not strings. This opens up
implementations for serialization/deserialization of binary formats.