Remove build script
I am not 100% sold on the idea of generating the manpages as part of the
release profile and using a build script. This might be better
implemented and served as a more generic build script for automating the
creation of releases or Continuous Integration (CI) stuff. Also, the
destination for the output of the manpage build is still at question.