Change `--framed` to `--sized`
The `--framed-input` and `--framed-output` flags are changed to
`--sized-input` and `--sized-output` to indicating the framing is by
data size. Other forms of framing, specifically by a delimiter byte, are
in the process of being added. Changing the names of the flags will
provide a more reasonable and predictable CLI for selecting different
forms of framing.