Add reading multiple files
Paths to multiple files can be specified as arguments to the
application. Each file is read, deserialized, serialized, and then
written. Unless the `-f,--from` option is specified, the file extension
for each file is used to determine the input format for deserialization.
If the `-f,--from` option is used, then all files will be deserialized
using the same format regardless of the file's extension.