Add colorized output of error messages
An error statement is written to `stderr` and it is composed to two
elements: (i) a tag and (ii) a message, which are separated by a colon
(`:`). The tag is on the left and the message is on the right of the
separator. The tag is now colorized with a bright red color if the
`stderr` is not piped or redirected, i.e. it is tty. This resolves Issue
6.